requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.1.7
docopt==0.6.2
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==5.4.1
pipreqs==0.4.9
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.15.1
whitenoise==4.1.2
yarg==0.1.9
gunicorn
